refactor(internal/python): add `generate_test_src` function from lines in `__main__.py`
