+++ internal/python/lib/generate_test_src/generators/__init__.py
@@ -1,0 +1,1 @@
+from .test_src import generate_test_src
